# [Spike] sdk/log: Change FilterProcessor to Filterer