--- a/sdk/log/logger.go
+++ b/sdk/log/logger.go
@@ -11,7 +11,6 @@ import (
 	"go.opentelemetry.io/otel/log"
 	"go.opentelemetry.io/otel/log/embedded"
 	"go.opentelemetry.io/otel/sdk/instrumentation"
-	"go.opentelemetry.io/otel/sdk/log/internal/x"
 	"go.opentelemetry.io/otel/trace"
 )
 
@@ -35,6 +34,12 @@ func newLogger(p *LoggerProvider, scope instrumentation.Scope) *logger {
 }
 
 func (l *logger) Emit(ctx context.Context, r log.Record) {
+	param := log.EnabledParameters{}
+	param.SetSeverity(r.Severity())
+	if !l.Enabled(ctx, param) {
+		return
+	}
+
 	newRecord := l.newRecord(ctx, r)
 	for _, p := range l.provider.processors {
 		if err := p.OnEmit(ctx, &newRecord); err != nil {
@@ -43,30 +48,15 @@ func (l *logger) Emit(ctx context.Context, r log.Record) {
 	}
 }
 
-// Enabled returns true if at least one Processor held by the LoggerProvider
-// that created the logger will process param for the provided context and param.
-//
-// If it is not possible to definitively determine the param will be
-// processed, true will be returned by default. A value of false will only be
-// returned if it can be positively verified that no Processor will process.
+// Enabled returns false if at least one Filterer held by the LoggerProvider
+// that created the logger will return false for the provided context and param.
 func (l *logger) Enabled(ctx context.Context, param log.EnabledParameters) bool {
-	fltrs := l.provider.filterProcessors()
-	// If there are more Processors than FilterProcessors we cannot be sure
-	// that all Processors will drop the record. Therefore, return true.
-	//
-	// If all Processors are FilterProcessors, check if any is enabled.
-	return len(l.provider.processors) > len(fltrs) || anyEnabled(ctx, param, fltrs)
-}
-
-func anyEnabled(ctx context.Context, param log.EnabledParameters, fltrs []x.FilterProcessor) bool {
-	for _, f := range fltrs {
-		if f.Enabled(ctx, param) {
-			// At least one Processor will process the Record.
-			return true
+	for _, flt := range l.provider.filterers {
+		if !flt.Filter(ctx, param) {
+			return false
 		}
 	}
-	// No Processor will process the record
-	return false
+	return true
 }
 
 func (l *logger) newRecord(ctx context.Context, r log.Record) Record {