@pellared 2024-09-18
@open-telemetry/go-approvers, PTAL. WDYT about such proposal? I am open to any/other suggestions.