@pellared 2024-09-19
> But I find this naming, `Filterer,` a little strange, as I am not sure there is a word for that. Maybe `Enabler`? 🤔

Initially I wanted to name it `Enabler`. However, it felt to me like it would be something that enables or disables a logger. Where it actually is an abstraction responsible for filtering log records. Moreover it is used in both `Enabled` and `Emit` handling.