@pellared 2024-12-18
Another feedback received during calls is that this proposal i less flexible and makes declaring some pipelines harder. 

For example, how to define 2 log record processing pipelines with different filters? For such case the filterer would have to be attached to a processor.

Chaining filters that are coupled to emitted logs is also awkward e.g. 

```go
etw := &ETWLogProcessor{} // it also has a "filter"
proc = &RateLimitingProcessor{
  Processor: etw,
  Duration: time.Second,
  Size: 1000,
} / it also has a "filter"
provider := log.NewLoggerProvider(
  log.WithProcessor(proc),
  log.WithFilterer(etw),
  log.WithFilterer(proc),
)
```